refactor(vnext): share bounded lexical primitives
